+++ src/admin_ui/src/MappingField.tsx
@@ -5,19 +5,19 @@ import FormLabel from '@material-ui/core/FormLabel';
 import FormControl from '@material-ui/core/FormControl';
 import {Labeled} from 'react-admin';
 import {makeStyles} from '@material-ui/core/styles';
-import {Theme} from '@material-ui/core/styles/createMuiTheme';
+import {Theme} from '@material-ui/core';
 
 
 const useStyles = makeStyles((theme: Theme) => ({
-  fieldset: {
-      padding: '5px 12px',
-      border: `1px solid ${theme.palette.grey[500]}`,
-  },
-  legend: {
-      paddingLeft: '5px',
-      paddingRight: '5px',
-      fontSize: theme.typography.pxToRem(12),
-  },
+    fieldset: {
+        padding: '5px 12px',
+        border: `1px solid ${theme.palette.grey[500]}`,
+    },
+    legend: {
+        paddingLeft: '5px',
+        paddingRight: '5px',
+        fontSize: theme.typography.pxToRem(12),
+    },
 }));
 
 export interface MappingFieldProps {